Bug fix for fuzzy matching capturing too much
It used to be that the regex 'A+B' would capture 'AA' even if the crossword was
a 2x2. This was because we were enforcing a fuzzy match without asking how long
the string could be.

This fix adds an optional arg to is_fuzzy_match to give the maximum length of
the string to match. This way, if the string given is the maximum length, the
entire regex must be matched, it cannot be fuzzy.